# docs/swagger.yaml
basePath: /
definitions:
  handlers.CreateUserRequest:
    properties:
      bio:
        type: string
      email:
        type: string
      password:
        type: string
      profile_picture_url:
        type: string
      username:
        type: string
    required:
    - email
    - password
    - username
    type: object
...
  models.User:
    properties:
      bio:
        type: string
      created_at:
        type: string
      email:
        type: string
      password_hash:
        type: string
      profile_picture_url:
        type: string
      role:
        type: string
      user_id:
        type: string
      username:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: API for OSOHUB using Cassandra
  title: OSOHUB API
  version: "1.0"
paths:
...
  /users:
    post:
      consumes:
      - application/json
      parameters:
      - description: User data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties: true
            type: object
      summary: Create a new user
      tags:
      - Auth & Users
  /users/{user_id}:
    get:
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      summary: Get user by ID
      tags:
      - Auth & Users
  /users/{user_id}/ban:
    patch:
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      - description: Ban (true) or unban (false)
        in: query
        name: banned
        required: true
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      summary: Ban or unban a user
      tags:
      - Auth & Users
...
  /users/me:
    get:
      description: Get the profile information of the currently authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get current authenticated user info
      tags:
      - Auth & Users
    patch:
      consumes:
      - multipart/form-data
      description: Allows the user to update their username, bio, profile picture,
        and/or password
      parameters:
      - description: New username
        in: formData
        name: username
        type: string
      - description: New bio
        in: formData
        name: bio
        type: string
      - description: New profile picture (JPG, PNG, WebP, max 10MB)
        in: formData
        name: profile_picture
        type: file
      - description: New password
        in: formData
        name: password
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update the authenticated user's profile
      tags:
      - Auth & Users
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...